+++ Tests/Recurly/Subscription_Test.php
@@ -16,15 +16,15 @@ class Recurly_SubscriptionTest extends Recurly_TestCase
 
     $add_on = $subscription->subscription_add_ons[0];
     $this->assertInstanceOf('Recurly_SubscriptionAddOn', $add_on);
-    $this->assertEquals($add_on->name, 'IP Addresses');
-    $this->assertEquals($add_on->add_on_code, 'ipaddresses');
-    $this->assertEquals($add_on->unit_amount_in_cents, 200);
-    $this->assertEquals($add_on->quantity, 2);
-    $this->assertEquals($subscription->collection_method, 'manual');
-    $this->assertEquals($subscription->po_number, '1000');
-    $this->assertEquals($subscription->net_terms, 10);
-    $this->assertEquals($subscription->tax_in_cents, 0);
-    $this->assertEquals($subscription->tax_type, 'usst');
+    $this->assertEquals('IP Addresses', $add_on->name);
+    $this->assertEquals('ipaddresses', $add_on->add_on_code);
+    $this->assertEquals(200, $add_on->unit_amount_in_cents);
+    $this->assertEquals(2, $add_on->quantity);
+    $this->assertEquals('manual', $subscription->collection_method);
+    $this->assertEquals('1000', $subscription->po_number);
+    $this->assertEquals(10, $subscription->net_terms);
+    $this->assertEquals(0, $subscription->tax_in_cents);
+    $this->assertEquals('usst', $subscription->tax_type);
 
     # TODO: Should test the rest of the parsing.
   }
@@ -134,5 +134,45 @@ class Recurly_SubscriptionTest extends Recurly_TestCase
       "<?xml version=\"1.0\"?>\n<subscription><account><account_code>account_code</account_code><billing_info><token_id>abc123</token_id></billing_info><address></address></account><plan_code>gold</plan_code><quantity>1</quantity><currency>USD</currency><subscription_add_ons></subscription_add_ons></subscription>\n",
       $subscription->xml()
     );
+
+  /**
+   * @expectedException Recurly_Error
+   */
+  public function testNonPreviewableSubscriptions() {
+    $this->client->addResponse('GET', '/subscriptions/012345678901234567890123456789ab', 'subscriptions/show-200.xml');
+
+    $subscription = Recurly_Subscription::get('012345678901234567890123456789ab', $this->client);
+    $subscription->preview();
+  }
+
+  public function testPreviewableSubscription() {
+    $this->client->addResponse('POST', '/subscriptions/preview', 'subscriptions/preview-200.xml');
+
+    $subscription = new Recurly_Subscription(null, $this->client);
+    $subscription->plan_code = 'gold';
+    $subscription->currency = 'USD';
+    $subscription->net_terms = 10;
+    $subscription->collection_method = 'manual';
+    $subscription->po_number = '1000';
+
+    $account = new Recurly_Account();
+    $account->account_code = '123';
+
+    $subscription->account = $account;
+
+    $subscription->preview();
+
+    $this->assertEquals(10000, $subscription->cost_in_cents);
+    $this->assertEquals('pending', $subscription->state);
+  }
+
+  /**
+   * @expectedException Recurly_Error
+   */
+  public function testNonPreviewableSubscriptions() {
+    $this->client->addResponse('GET', '/subscriptions/012345678901234567890123456789ab', 'subscriptions/show-200.xml');
+
+    $subscription = Recurly_Subscription::get('012345678901234567890123456789ab', $this->client);
+    $subscription->preview();
   }
 }