+++ lib/recurly/subscription.php
@@ -10,13 +10,14 @@ class Recurly_Subscription extends Recurly_Resource
     Recurly_Subscription::$_writeableAttributes = array(
       'account','plan_code','coupon_code','unit_amount_in_cents','quantity',
       'currency','starts_at','trial_ends_at','total_billing_cycles', 'first_renewal_date',
-      'timeframe', 'subscription_add_ons', 'net_terms', 'po_number', 'collection_method'
+      'timeframe', 'subscription_add_ons', 'net_terms', 'po_number', 'collection_method',
+      'cost_in_cents'
     );
     Recurly_Subscription::$_nestedAttributes = array('account', 'subscription_add_ons');
   }
 
-  public function __construct() {
-    parent::__construct();
+  public function __construct($href = null, $client = null) {
+    parent::__construct($href, $client);
     $this->subscription_add_ons = array();
   }
 
@@ -26,6 +27,14 @@ class Recurly_Subscription extends Recurly_Resource
 
   public function create() {
     $this->_save(Recurly_Client::POST, Recurly_Client::PATH_SUBSCRIPTIONS);
+  }
+
+  public function preview() {
+    if (!empty($this->uuid)) {
+      throw new Recurly_Error('Cannot preview an existing subscription');
+    }
+
+    $this->_save(Recurly_Client::POST, Recurly_Client::PATH_SUBSCRIPTIONS . '/preview');
   }
 
   /**